First pass at log support using boost::log. This has a ridiculously large number of headers and tweaks to it; for now, we just log to the console, but can add a sink to log to files as well.
To use it, each program needs to call the set_log_level() function
which will wrap up all of the log setup; this means that you only
have to import log.h. The level itself is just a stub for now,
but will eventually be used to set priority in the config file.


git-svn-id: https://subversion.pnnl.gov/svn/NIMS/trunk@106 e0feaedf-03d4-4201-b3fb-2d10bec3ea31