Add support for NIMS_HOME environment variable. All applications are now launched relative to this variable, and it's set in the init script. If a user runs it from the shell, nims will exit if the variable has not been set.
As part of debugging this, I also fixed the SIGHUP behavior of
nims, so it now ignores SIGCHLD while manually stopping them
in a warm restart. This was causing an exit if you sent SIGHUP
too many times.

Additionally, I added checks for SIGINT to ingester, so the
interrupted syscalls are handled in both loops. Previously,
it was nearly impossible to exit cleanly with a SIGINT if
it was blocking on one of the early connect calls.


git-svn-id: https://subversion.pnnl.gov/svn/NIMS/trunk@231 e0feaedf-03d4-4201-b3fb-2d10bec3ea31